1_關鍵偵擷取/requirements.txt
tensorflow==2.15.0
ffmpeg-python
numpy
pandas
tqdm
protobuf<=3.20.3
ultralytics
lapx
opencv-python
